Move tests of CuDNN RNN weight conversion to a more proper place.
- move from cudnn_recurrent_test.py to test_model_saving.py
- adjust imports to be consistent
